pyproject.toml
[tool.poetry]
name = "Messy"
version = "0.1.0"
description = ""
authors = ["Alexander Viklund <bullfest@sthlm.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "~=3.1.6"
djangorestframework = "~=3.12.2"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-django = "*"
black = "^20.8b1"
pytest-factoryboy = "^2.1.0"
pytest-randomly = "^3.5.0"
pytest-freezegun = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rs"
DJANGO_SETTINGS_MODULE = "messy.settings"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]
python_classes = [
    "Test*",
    "*Test",
]

